make use-socket hook look better
