+++ apps/client/package.json
@@ -68,7 +68,7 @@
     "prettier": "^2.8.0",
     "prettier-plugin-tailwindcss": "^0.1.13",
     "tailwindcss": "^3.2.2",
-    "ts-jest": "^29.0.3",
+    "ts-jest": "^29.1.0",
     "type-fest": "^3.5.1",
     "typescript": "^5.0.4",
     "webpack": "^5.75.0"
