+++ apps/client/src/lib/hooks/use-socket/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./hook"
